refactor(backend): 更新核心模块和服务
更新认证、路由、服务、工作器和数据库操作等后端核心组件。

+++ packages/backend-rust/src/routes/optimization.rs
@@ -610,7 +610,7 @@ async fn evaluate_recent_learning_effect_pg(pool: &sqlx::PgPool) -> Result<Optio
         SELECT
             COUNT(*) as "total",
             SUM(CASE WHEN "isCorrect" THEN 1 ELSE 0 END) as "correct",
-            AVG(NULLIF("responseTime", 0)) as "avg_response_time"
+            AVG(NULLIF("responseTime", 0))::float8 as "avg_response_time"
         FROM "answer_records"
         WHERE "timestamp" >= $1
         "#,
